Refactor serialization initialization and kryo usage. StreamingManager has better security wrt session IDs

+++ src/dorkbox/network/connection/streaming/StreamingControl.kt
@@ -1,5 +1,5 @@
 package dorkbox.network.connection.streaming
 
-data class StreamingControl(val state: StreamingState, val streamId: Long,
+data class StreamingControl(val state: StreamingState, val streamId: Int,
                             val totalSize: Long = 0L,
                             val isFile: Boolean = false, val fileName: String = ""): StreamingMessage
